Create LICENSE
I came across this project through the msextra thread where you documented this work: https://www.msextra.com/forums/viewtopic.php?f=67&t=52539

If your intention was to share this code as open source, it is good practice to add a software license to allow others to use this. I've selected a basic MIT license to add to this repo, feel free to update to another as desired.